+++ integrator.h
@@ -18,20 +18,20 @@ namespace integrators
 	{
 		protected:
 			//Delta t value used.
-			float timeStep;
+			double timeStep;
 			//Keeps track of total time.
-			float systemTime;
+			double systemTime;
 
 		public:
 			//Generic functions for advancing position and velocity
-			virtual int nextSystem(float index, mathTools::points* pnt, physics::forces* f)=0;
+			virtual int nextSystem(double index, mathTools::points* pnt, physics::forces* f)=0;
 
 			//getter/setter for 'timeStep'
-			void setTimeStep(float newStep) { timeStep = newStep; }
-			float getTimeStep() { return timeStep; }
+			void setTimeStep(double newStep) { timeStep = newStep; }
+			double getTimeStep() { return timeStep; }
 
 			//getter for 'systemTime'
-			float getSystemTime() { return systemTime; }
+			double getSystemTime() { return systemTime; }
 			//output current system time.
 			void writeSystemTime() { std::cout << "Current time: " << systemTime << ".\n"; }
 			//increases system time by one timestep.
@@ -48,19 +48,19 @@ namespace integrators
 		public:
 
 			//Constructor/Destructor
-			verlet(float dt);
+			verlet(double dt);
 			~verlet();
 
 			//Calculates the next system state.
-			int nextSystem(float index, mathTools::points* pnt, physics::forces* f);
+			int nextSystem(double index, mathTools::points* pnt, physics::forces* f);
 
 			//Gets the next position and velocity terms.
-			int nextPosition(float index, mathTools::points* pnt, physics::forces* f);
-			int nextVelocity(float index, mathTools::points* pnt, physics::forces* f);
+			int nextPosition(double index, mathTools::points* pnt, physics::forces* f);
+			int nextVelocity(double index, mathTools::points* pnt, physics::forces* f);
 
 			//The velocity verlet Algorithms to obtain the next position and velocity terms.
-			float posAlgorithm(float pos, float vel, float acc, float t);
-			float velAlgorithm(float pos, float vel, float acc, float accnext, float t);
+			double posAlgorithm(double pos, double vel, double acc, double t);
+			double velAlgorithm(double pos, double vel, double acc, double accnext, double t);
 	};
 
 }
